core: introduce sandboxing and node-hash caching
With this changeset, build rules will only run within a sandbox where
only the declared inputs are directly available.

The process will also cache rule outputs based on the hash of the rule,
which includes the hashes of all of the inputs.

Consequently, the caching code has been greatly simplified.