core: rule srcs/outs as function of dep. artifacts
This change begins the path to allow a rules' inputs and outputs to be
determined based on the artifacts of their dependencies.

This allows us to build a much richer dependency graph where certain
build nodes will produce outputs that will be taken as vanilla input
sources by other rules, and thus be used to declare additional, dynamic
outputs.

Take for example the ErlangLibrary rule. Initially it only could declare
outputs based on the inputs that were gathered during the crane file
evaluation time.

If you had a `srcs = ["hello.erl"]` that meant that that you could only
meaningfully declare an output for `"hello.beam"` based on this source.

However, a rule could _create_ an Erlang source file, which would then
be useful for an `ErlangLibrary` rule to consume.

So our new rule, in this case CaramelLibrary, declares an artifact that
is in fact a `.erl` source files.

The ErlangLibrary rule will now be fed the declared artifacts of its
dependencies, and it can filter out the things that it cares about, such
as other `.erl` source files.

This isn't implemented for all rules yet, but the plumbing is there.