core: cache computed srcs/outs/hash in BuildNode
Normally on every new node that we pick up, we try to compute its hash
to check for cache hits.

This means that we need to compute its final declared inputs, and
declared outputs.

Since the last change, the declared inputs and outputs are also a
function of the artifacts of the dependencices of our node.

this forces us to to traverse the dependency graph upwards from our
node, finding the transitive dependency artifacts until the end of the
graph, to then come all the way back and compute our own declared inputs
and outputs based on it. This is because our dependencies also need
those artifacts to compute their final inputs/outputs.

This changeset will cache the final inputs and outputs declared for
every rule at the build node level, just like we saved the hash of a
node.

That way we just have to go one level up to get all the relevant
information we need, on whatever point in the graph we are.