core: add prelim. support for lumen binaries
Lunch break!

Here's a thought: bootstrapping.

Everything I'm about to write is super hand-wavy, but it _is there_.

We've got the pipelining in place to build large projects. We got a
release of Caramel that can be used to type-check _a small subset of
Erlang_. We got the OCaml compiler support for binary compilation.

What's stopping Crane from orchestrating a build of itself in
type-checked Erlang that compiles to a single binary?

It hit me when integrating Lumen, that the binary we get out would be as
convenient to a user as the tool used to build it: Crane. So why not
have Crane build itself and slowly drop the dependency on the Rust
ecosystem?

This would make the project a lot easy to get into and contribute, and
would allow us to write it in a mix of Erlang, OCaml, Gleam, Clojerl,
etc. A perfect test-bed for language interoperability tests.

Off the top of my head, we could start by using Caramel if we
supported a larger standard Erlang, but unless we start running
Multi-core OCaml at the bottom layer, we won't be able to do parallel builds in
the current model.

If that works, then we could try to build it also with Lumen, which
AFAICT would support the parallelism in the same way the ERTS would.

This would allow Crane to bootstrap itself.