fix(core): better surface errors when locally caching things
A small issue we had when the latency of the remote cache grew _a bunch_
is that dependent targets just started to build before the targets had
been properly saved in the local cache.

Now we first save to the local cache before trying to replicate to the
remote cache. This allows us to just spin up the Future and keep going
with the build.