feat(core): support relative labels in subtrees
Relative labels in subtrees allow us to make a target depend on another
target that is _contained_ within the same directory subtree.

For example, if `//a:lib` depends on `//a/b:lib`, it can write it out
as:

```toml
[[rule]]
name = "rule"
deps = [ "./b:lib" ]
```

Which helps make target `a` relocatable, to other parts of the
repository without having to juggle around the subtarget paths too much.

This, in conjunction with some form of "visibility rules", can help make
a monorepo much easier to refactor/maintain over time by preventing
targets from depending on deeply nested targets.