feat(core): smarter checks in file scanner
Now we first check if the file matches what we are looking for, and bail
out early if we do.

If it doesn't, then we already know we don't care about this file so we
only have to check if it is a directory.

If it *is* a directory, it may still be part of the directories we want
to ignore, so we check for that to make sure we don't queue dirs that we
want to avoid.