fix: simplified label resolution
From now on, any label will always have a path that is relative to the
root of the workspace the label belongs to.

This means that relocating labels has no logic, and it makes it easier
to send a label off to a remote workspace (or a remote worker) and build
it there.

ALL label paths are then of the form `./path/to/package` -- including
the ones from URLs. These are formatted accordingly when we parse the
URL path (see `LabelBuilder::from_url/2`).

The additional flag that local labels have, `kind`, is set in the
process of cleaning up the paths, and it is used when reconstructing
labels that are relative to a Build file.

For example, if the buildfile `hello/Build.toml` has these targets:

```toml
[[erlang_library]]
name = "supporting_lib"

[[erlang_library]]
name = "main_lib"
deps = [":supporting_lib"]

[[erlang_script]]
name = "main"
deps = [":main_lib", "./vendored/lib"]
```

Then the label struct for `":supporting_lib"` and `"./vendored/lib"`
will be marked as Relative on construction, which lets the Target
building logic reconstruct them to have the proper paths:

* `:supporting_lib` becomes `./hello:supporting_lib`
* `./vendored/lib` becomes `./hello/vendored/lib`