feat: reuse remote workspaces cache across workspaces
In this change we make every remote workspace have its own store prefix,
rather than share the store prefix of the workspace that depends on it.

This makes it possible to share the cache of remote workspaces across
any workspace that depends on it.

For example, if you had workspace A (with prefix `/warp/store/1-A`) and
wanted to build `https://remote.workspace/target` (with hash `1234`),
the final store key would ahve become: `/warp/store/1-A/1234`.

You can see how the hash of the target, `1234`, was saved in the cache
of the workspace that _required it_, rather than the cache of the remote
workspace itself.

Now with this change, the workspace `https://remote.workspace` gets a
store prefix (say `/warp/store/2-remote.workspace`) so the contents of
its targets will be cached there.

This way `https://remote.workspace/target` ends up in `/warp/store/2-remote.workspace/1234`.