refactor: support building multiple labels
Since the entire infrastructure of the BuildExecutor supports queueing
and building labels parallely, it only makes sense that we should be
able to start a build with more than one label, instead of waiting one
label at a time for things to build.

This refactor brings support for this, and allows the status reporter to
print out a Finished line with relevant formatting when we have more
than one goal label.