refactor: unify serde for label to support naive deser from string repr
This means that if we try to deserialize something that is structured,
we'll assume it has all the fields we need to put together a full Label
(with a `workspace`, a `name`, and a `inner_label`).

However, if we just get a string, like `"https://warp.build/tool/thing"`
then we will parse that correctly instead of blowing up.