refactor: introduce warp engine
The Warp Engine is an abstraction on top of the entire core crate, to
help us minimize the boilerplate used in every binary command.

It also ensures that the interface for interacting with Warp as a system
is homogeneous:

1. WarpEngine's builder chain
2. initialize
3. execute
4. shutdown

In the future this will help stack a gRPC API on top of it to call
remote Warp Engines from the CLI as well.