feat: massive improvement in status reporter latency
This commit introduces a new command called `hash`, that we are using to
explore the theoretical speed limits at which Warp can operate.

Through that work, I discovered that when we simply read and hash all
the sources in a workspace (from small [1-50 files] to massive [20,000
files]), the largest time was always spent consuming the messages in the
StatusReporter.

The StatusReporter now enjoys its own execution thread, which guarantees
that it will only be a few milliseconds behind the Workers, and so if
you run `time warp` you will see that time we report for a build is only
a dozen milliseconds lower than what `time` sees.

Additionally, some other changes we've made are:

1. a new Queuer Worker spins up and queues up all targets in a workspace
   concurrently with the rest of the Build Workers -- this guarantees
   that we can start working as soon as a single target has been queued.

2. the Main Worker runs within the LocalHandlerPool, just like all the
   Helper Workers -- this guarantees that the main thread of the
   application has plenty of time to poll all the futures instead of
   accidentally being caught up in some blocking task.