feat(beam): make dependency profiles configurable
This helps us avoid certain dependency cycles that some libraries have
with the tools that use them.

For example, EMQTT depends on a subset of EMQX for testing purposes.
This makes sense since you'd want to bring up an EMQX node to be able to
test EMQTT against it.

However, EMQTT is also a *dependency of* EMQX.

So if we naively pull all dependencies, we will end up cycling:

* `emqx` has a production dependency on `emqtt`
* which has a testing dependency on `emqx`

So splitting these and being able to pull in `emqtt` _without_ pulling
its transitive testing dependencies, we break the cycle.

This works because we are *not* testing EMQTT when we test EMQX. The
opposite isn't true, and if EMQTT was the main repo, we'd have to bring
in EMQX, and the cycle woudl still be there. We'll solve that when we
get there.