feat(core): get mark II's worker pool infra in a testable state
The bulk of the work that we are doing here is re-structuring mark I
code into subsystems so that things are tidy for extension and better
testing.

This first commit focused on getting the WorkerPool to compile and test,
and it includes:

* a new `SharedContext` to avoid passing dozens of references around
* the `LocalWorker` (without the main loop, and planning components)
* a change from `Label` to `Target`
* unifying on `Signature`s (so the old `Target` is gone)
* the `TaskQueue` and `TaskResults`
* and more

The way the codebase is structured prevents us from having a single
top-level `models` module, so instead we'll define models within their
relevant phases. Eg, `Signature` is defined in the `resolver` stage, and
`TargetManifest` in the `executor` stage.

More coming soon!