test(MarkII): property-test the event channel
We are also preparing this module to be wire-friendly by making the
Event datatype contain only primitives and no reference data (like types
implementing the Error trait, which aren't clonable).

In the future this data will be fit into protobuf messages that we can
use in a remote log-streamer protocol.