test(MarkII): first concurrent tests
We're starting to do randomized concurrent testing of the shared-memory
components that will be access by many threads and have some form of
locking in them.

This was our first time using `shuttle`, so we'll improve over time.

In this commit we're testing that consumption of the Task Queue from
multiple threads doesn't send the same task to multiple threads at the
same time. We need to add more tests for this behavior (such as what
happens during next/nack cycles, and how does the internal `wait_queue`
behave when we use `queue_deps`).

The other thing we're testing here is that we _always_ serialize the
registration of targets, to make sure that the same target registered
twice by two different threads always yields the same target id. This
prevents us from having the same target being considered two different
things.