feat(MarkII): introduce worker::Context trait
In order to test the WorkerPool, we need to provision mock workers that
can be used to test specific behaviors. All workers share a Context,
but any given test may have specific mock workers that need a specific
mock context to really test something.

For this purpose we introduce the Context trait, which is implemented by
different structs (such as `LocalWorkerContext` or `FixtureContext`).
This trait is expected to be defined as an associated type of a type
that implements the `Worker` trait. This allows us to say that a given
type that is a Worker has a fixed specific context type associated to
it.

The only requirement on the types that implemnet `Context` is that they
should provide a way to extract _results_ from it.