feat(MarkII): introduce Resolver trait
In order to test the LocalWorker, which depends on a specific Resolution
Strategy (now an `impl Resolver`), we need to be able to mock different
resolvers.

To do this we are introducing the `Resolver` trait, which can then be
used to parametrize the `LocalSharedContext`. This allows us to create
different contexts that change the behavior of a local worker, so we can
test for different scenarios like:

* What happens when the Resolver _fails_
* What happens when the Resolver determines we need to prepare other
  dependencies first
* etc

Before we were bound to testing the actual resolver, and now we can just
feed the right one into the context.