feat(MarkII): introduce and test PublicStore
The main change is the introduction of the PublicStore within the flow
of the DefaultStore. This PublicStore encapsulates the basic logic to
download and install public packages from
`https://public.store.warp.build`, which is a new piece of
infrastructure backed by AWS CloudFront (so it'll be hella cheap and
fast).

Other than that, in this changeset we are sketching the entire flow from
the DefaultResolver downwards, meaning that we introduce:

* The PublishManifest -- a structure representing a published package in
  the PublicStore
* The Store trait -- to ease testing of multiple stores
* The DefaultStore -- to orchestrate the default store configurations
  (public, local, and in the future a shared/private one)
* The PublicStore -- described above
* The TricorderManager -- finds, downloads, and runs tricorders
* The TricorderRegistry -- to abstract how we find / update tricorder versions.
* The GrpcTricorder -- an implementation

On top of this we updated all dependencies and we're cleaning up other
code as we go along:

* The `rules/config` module now is just a tiny language without all the
  protobuf serialization mess

* The ArchiveManager only includes a download function, instead of all
  the other ones we used to have

* The Config struct now holds the HttpClient pool used by everything in
  the system

* The Config struct has a configurable `store_root` that can be used for
  testing purposes with the `assert_fs` library to change the path of
  were artifacts are stored. This pattern will allow us to test
  file-system side-effects with ease.

That's all folks. We continue making the above pieces testable so we can
move fast and break nothing.