feat(MarkII): tricorder manager can install, spawn, connect, and ready tricorders
The main change here extends and tests the TricorderManager to be a
thread-safe way of finding, installing, starting, and readying
tricorder services.

All workers will share a TricorderManager and will get back an `impl
Tricorder` struct that they will own individually. This allows for every
thread to talk to the server independently.

We've also tested 4 important cases for the manager:
1. it finds and install a tricorder
2. it fails loudly if the tricorder can't be found
3. it fails loudly if the tricorder can't be started
4. it fails loudly if the tricorder can't be readied

Other changes:

* PublicStore now has split CDN and Metadata urls. Typically cdn will be
  `public.store.warp.build` and point to CloudFront, and metadata will
  be `store.warp.build` and point to a Github Pages repository where
  updates happen for free and faster.

* TricorderRegistry now uses the configuration's PublicStore Metadata
  URL as a default.

* ArtifactManifest now can build its own Id, and find specific provided
  binaries.

* DefaultStore can canonicalize an artifact provided by a manifest.

* Brought back the PortFinder