feat(MarkII): first working version of MarkII 🚀
This changeset makes MarkII _execute and locally cache_ results from the
resolution/planning phase. In other words, let there be warp!

We threaded in the WorkspaceManager to make sure that when we're
concretizing targets we use relative paths that can be relocated to the
Store for execution.

We're also buidling the ArtifactManifest straight out of the results of
the execution and validation, and fixed a couple of bugs along the way.

The ArtifactManifest is going to drop a few fields like `cached` and
`is_valid`.