feat(core): introduce network resolver
This changeset restructures the DefaultResolver to make use of several
smaller resolvers depending on configuration, goal, and target kind.

In particular we are introducing the NetResolver to help us do
resolution of targets that are URLs. It relies on the ArchiveManager to
download and extract things, on the TricorderManager to find the
appropriate tricorder, and on the Tricorder itself to _prepare_ a
dependency after it has been stored.

We are also introducing new configuration (the `remote_workspace_root`
path), extending the URL requirement to include the URL of the tricorder
that should handle it, and cleaning things up as we go along.