feat(schemas): new fields for better dep resolution
We extend the UrlRequirement message to include a subpath. This can be
used to specify what subpath of a specific url is relevant as a
requirement. For example, if the URL is a Git repository, this can help
us do subtree clones. If the URL is a tarball, this is used to narrow in
the uncompressed output, etc.

The Dependency message now includes a list of all the outputs that
Dependency has generated. This makes it easier to write rules that rely
on paths for includes.

The PrepareDependency RPC now also includes the name of the package that
is being prepared.

We also regenerate all relevant code.