feat(core): carry over requirements as dependencies
This changeset implements 2 flows:

1. preparing a dependency from a URL requirement
2. carrying over requirements as dependencies

Both flows we had in the Mark I.

Since the introduction of the NetResolver, we've been able to receive a
UrlRequirement and download/unarchive it as needed. A missing part here
was to know what to do with the unarchived URL.

To handle this we reintroduce the PrepareDependency RPC on the Tricorder
service. This RPC takes care of setting up a folder locally so the
unarchived URL is usable by warp.

Sometimes (like with hex.pm packages), the original unarchiving still
has an inner archive (`contents.tar.gz`) that needs to be unpacked and
handled by the Tricorder instead. To know exactly what name the final
subpath within the archvie has, we make use of the `subpath` field
introduced before.

Once a Tricorder reports missing requirements for a particular target, we keep
those dependencies around and pass them back into the tricorder
accordingly.

Then one the signature is generated, we still include those original
dependencies as part of the set of dependencies that are an input to the
rule.

This means that Tricorders have to do a little less work keeping track
of those originally missing requirements.

To do this we:

* keep a set of dependencies in the TaskResults, instead of replacing it
  every time we requeue dependencies

* we fetch all these dependencies before passing them on to the
  Tricorder rpcs

---

More so, we make some changes in the way we compute the code paths and
include paths in the BEAM tricorder.