feat(tricorders/rust): test matcher for signature
This commit uses the newly created rust_test rule. It temporary
replaces the use of any other rules in favor of this one for testing
purposes.

This also replaces the use of any Symbol received from warp, since that
is no longer used. This was replaced by the TestMatcher, since this was
only specific to tests.

For testing all of these, I have also introduced `p003` as a new test
project.