feat(core): prepare events for the new status reporter
This change makes it easier than ever to send events from anywhere in
the app by carrying around in our Warp Config object a pointer to the
Event Channel.

We also did some renamespacing of Events, so they are easier to manage, and
easier to use. Queue events are in the `QueueEvents` enum, etc. Those
can be passed directly into the `EventChannel`.

I'm opting for this approach rather than a more generic `Event` trait,
to maintain as much type-information/structure as possible, to build
better status reporting on the other side.

These events, however, are wire-friendly. So remote builds should be
able to encode them into Protobuf messages with ease.