feat(cli): restore status reporter
This changeset restores the status reporter from the Mark I, albeit with
some modifications.

We have a `Reporter` trait that implements default handler functions
over the different `Event` types from `warp_core`.

The `StatusReporter` is now keeping internal state, so that the handler
functions can update it as things are happening.

More events to come!