feat(core): introduce TestMatcher
The new TestMatcher and TestMatcherId structs help us capture more user intent
when they are executing the `warp test` command.

Now this command will take a `target` as its first input, and it will
collect the rest of the arguments into a _test matcher_. This
TestMatcher will help the Tricorder generate the appropriate signature
to test only the tests that match it.