feat(core): support multiple signatures
This changeset allows us to handle multiple signatures coming from the
resolution phase, including maintaining references as a task gets
replaced by another due to the signature split.

For this to work, the unit of work has to go from a single Target, to a
compound key including the Target _and_ the Signature that is being
built.

Tasks were the clearest place where we could add the signature
information to make all tasks in the queue and in the results unique.

Since signatures are rather heavy objects, we introduce the
SignatureRegistry to keep track of our signatures while issuing
lightweight copyable handlers (SignatureIds).

Because some Tasks get replaced by others when they are split into
several new tasks due to the signature resolution, we needed also to be
able to _replace_ what a Task was, without needing to replace all of the
references to that task.

For that we introduced the TaskId handler, that is issued by the
TaskRegistry, which keeps track of all the tasks in place.

And lastly, to avoid anyone being able to build a Task that will go
unaccounted for or that will not be possible to _update_ through the
TaskRegistry, we introduced the UnregisteredTask struct.

This struct is used to capture the intent of creating a task, without
actually creating one.

Other than this, all the code has been updated to work with the `Task`
struct instead of receiving individual Goal, TargetId, and SignatureId
information.

fix(core): rule expander double-root append

The Rule Expander now makes sure that it does not add the root of a
Signature to the paths contained in the Rule Config IF the paths already
have that root as a prefix.

Example:
  * root=./a/b path=./a/b/file => ./a/b/file
  * root=./a/b path=./file     => ./a/b/file

feat(core): send goal to tricorders

In order to avoid filtering sigantures in the core, we will send the
current Goal to tricorders to let them figure out what are the necessary
signatures.