fix(core): hash sources deterministically
When comparing the artifact hash of the Rust tricorder with @capitanu,
we found out that they were diverging in his and my computer. Tracking
down this bug, we found that the sources were not being hashed in a
sorted manner (probably because the `FxHashSet` iterator doesn't
guarantee ordering), and then that the `SourceSet` was not normalizing
source paths, so sometimes we got `./` at the beginning of some relative
paths.

This changeset removes this nondeterminism, and makes us both build the
_exact_ same artifact.