feat(tricorders/beam): multiple subtrees
When generating a subtree for a test matcher, we will produce as many
subtrees as relevant test matches we find in the files. This allows us
to match the subtrees to signatures, similarly to what we did in Mark I,
and thus rebuild the AST-based caching we had before.

We also rewrote the `erl_tree_splitter.erl` module into Elixir, and made
sure to remove all line/position information from the AST so that new
lines on other tests do not affect the other generated subtrees.