feat(core): reintroduce AST-based caching for tests
We're also refactoring the codebase here to pass around a `CodeManager`
that can be used to perform diverse actions that are related to source
code. Most of these will rely on the CodeDatabase and the
TricorderManager.

Further caching is warranted here, especially an in-memory cache during
program execution, to make sure we only do I/O if we absolutely need to.

But generally speaking this approach means that we can reuse the
signature generation logic (which includes subtree chunking and
persisting them), from anywhere in the program that needs to do a fresh
(or cached) check of the contents of a file.