feat(core): pack command updates local manifests
Package now updates the local manifest.json files whenever we are using the
Pack Command with an FsTarget. There is currently no action being taken when
the Pack Command is passed a RemoteTarget or an Alias.

FsTargets are now normalized using the workspace root, so that we're always
working with an absolute path, even if the user passes a relative path. This is
also used in the Build Commands.

The store directory tricorder is renamed to tricorders.

Close #37

Other changes:
* fix(core): rename tricorder to tricorders
* fix(tests): upgrade to mockito 1.0.2
* fix(tests): test expects specific port to match hash
* fix(core): set workspace path for each worker
* fix(core): provide workspace path when updating manifest